Several modifications and bug fixes in a debugging session Fri 17.5.2018 Below the email that I sent as well:
Hi Lewis,

I had a closer look into the code. First of all, attached is a LOG file of the sequence of boards that you
are printing in your code at the moment. search for lines starting with
// HWL
for my comments on the scoreboards and the moves made. This are general observations on the behaviour,
and some things that you want to check.

I also looked into and made changes to the code. Likewise search for lines with
// HWL
or just the tag HWL in the code (mainly AIPlayer.cs).
There are several issues commented on, but the main is in line 1189:
            else if (ply > 0) // HWL: BUG: this doesn't look right: it seems you NEVER look-ahead to the next move!
	                {
			                // check
					                score = EvalCurrentBoard(board, scoreBoard, ourindex, us);  // is current pos a win?
							            }

In effect what you are doing here is to always finish the tree search at the current level, because in the
default settings ply starts with 2, and therefore you don't get a full tree search. You can check this by
enabling the ReadLine() in line 1203: this will never bit hit.
Without a proper tree search you just check for special cases on the current board, and you never look ahead,
so this will give poor behaviour. I changed the code, taking the else branch out, starting ply with 0 and setting
maxPly to 2 (larger values will probably give you a heap overflow because the search space is too large!).

Another important issue is that you pick a move in line 1234:
                     Move = availableMoves[i];
		     but you actually never use the move: you never put a piece there, just pass it tot he rec call , and at the start of the
		     call positions is immediately overwritten by a random move. I think I've fixed the immediate lines there.

In the modified version you now have a tree search and in line 1263 I print the main info; note that ply is printed,
showing how deep in the tree you are, verify that you do search the first 2 levels of the tree. I'm also attaching
a LLOG file with output from this run. Search for lines starting with
**HWL
to trace how the values change, which moves are tried etc.

The entire run still gives bad moves, but I think core issues are fixed now.

What I'd recommend to do, is to follow in detail such a trace for making just 1 move, on a constructed
board, where you can estimate where you find good moves (linking up 2 in a row, within 2 moves), and
bad moves, and check that the scoreboards that you are printing are as expected, and that the one move
returned is the right one.

If you run the entire game, you won't be able to trace how the scores are put together and you need this in
order to verify the Minimax is working properly

I'll push my changes into the repo (based on the 170519 checkin), and I also attache a tar file with the sources,
and the 2 LOG files that I mentioned.

I think with these changes you are on the right track to get the expected behaviour together. The important thing
is the full tree search. Don't worry too much about localised checks of patterns on the board. Once you have a
fully function tree search with plausible moves, you can test and compare this with a parallel version, and then
you have the core of a good thesis

Cheers,
--
 HaWo